==> 02-Combine-Tenure-MPI-Data.Rmd <==
---
title: "People-centric restoratrion: Optimizing human development variables and land tenure"
author: "Pooja Choksi"
date: '2022-06-03'
output: html_document
---

```{r setup, include=FALSE}
library(ggplot2)
library(forcats)
library(dplyr)
library(qcc)
library(gridExtra)
library(viridis)
#get color blind friendly colors 
library(pals)
```

PEOPLE-CENTRIC RESTORATION: MAXIMISING HUMAN DEVELOPMENT AND ADDRESSING VULNERABILITY

## QUESTION 1: 
Where are the districts which would provide the maximum benefit when considering the following variables: (1) max population, (2) Multidimensional poverty index and (3) climate vulnerability, (4) total number of hectares available for restoration? 

Working on the assumption that ecological restoration is beneficial for natural resources dependent livelihoods, we quantify people-centric restoration potential of different land tenure regimes and states to meet human development needs, using public land use data. 

In the following chunk of code we combine all data sets, which include:
1. Census 2011 population and land use data 
2. Multidimensional Poverty Index (Oxford, 2018)
3. Climate Vulnerability Index (National Comission on Climate Change Adaptation)
4. Number of households with job cards 
5. Number of households employed by MNREGA
6. Number of households with and without land 

```{r message= FALSE, warning= FALSE, include=TRUE}

#1. Land use/ de jure land tenure data from script 01
dat.summary = read.csv("All-State-LandUse-Combined.csv")

#2. Multidimensional Poverty Index from Oxford Poverty and Human Development Initiative
mpi_district = read.csv("Multidimensional_Poverty_Index_2018.csv")
mpi_district = mpi_district%>% drop_na(Living_standards)

```

AGGRGATE ALL DATASETS TO DISTRICT LEVEL

```{r message=FALSE, warning=FALSE, include=TRUE}

#1. Land use/ De jure land tenure data
district_summary = dat.summary%>%dplyr::group_by(District_Name, State)%>%summarise_at(.vars =vars(Total_Indiv, Total_Common, Total_Forest, Total_Unaccounted, Total_Population, Total_area_in_hectares), .funs =c(sum = "sum"))

#MPI data is available at the district level already
```

COMBINE LAND USE/ DE JURE LAND TENURE WITH MPI DATA

There are several mismatches in district name etc between the census 2011 data and the MPI data. In this chunk of code addresses the inconsistencies before combining the two datasets.

```{r message= FALE, warning= FALSE, include= FALSE}

#ensure names of districts and states match in mpi and dat.summary 
conflict_state = mpi_district[!mpi_district$State %in% dat.summary$State, ]

unique(conflict_state$State)

#change State names to match dat.summary 
dat.summary = dat.summary%>%dplyr::mutate(State = recode(State, "Andaman Nicobar UT" ="Andaman And Nicobar Islands","Daman Diu" ="Daman And Diu",
"NCT_Delhi" ="Delhi",
"Jammu Kashmir" ="Jammu And Kashmir",
"Dadra Nagarhaveli UT" ="Dadra And Nagar Haveli",
"Tamil_Nadu" ="Tamil Nadu",
"West_Bengal" ="West Bengal"))

#check again
conflict_state = mpi_district[!mpi_district$State %in% dat.summary$State, ]

unique(conflict_state$State)

#The states/ UTS that remain as conflict are those that do not have and land use data and were removed in script 01 - Meghalaya. Mizoram, Puducherry

#here we filter them out to remove from analysis
to_remove = mpi_district%>%filter(State == c("Meghalaya", "Mizoram", "Puducherry"))
mpi_district = anti_join(mpi_district, to_remove)

#Join mpi and dat.summary using state and district names
trial = merge(dat.summary, mpi_district[c("District", "State", "MPI_District", "Number_of_MPI_PeopleDistrict")], by.x = c("State", "District_Name"), by.y= c("State", "District"))

#find the districts with no match between mpi and dat.summary 
conflict_obs = mpi_district[!mpi_district$District %in% trial$District_Name, ]

#Recode MPI data district names to match dat.summary 
dat.summary = dat.summary%>%dplyr::mutate(District_Name = recode(District_Name, 
"Uttar_Bastar_Kanker" ="Uttar Bastar Kanker",
"Janjgir_Champa" ="Janjgir - Champa",
"Kabeerdham" ="Kabirdham",
"Koriya" ="Korea (Koriya)",
"North_East" ="North East",
"North_West" ="North West",
"South_West" ="South West",
"Panch Mahals" ="Panchmahal",
"Sabar Kantha" ="Sabarkantha",
"Banas Kantha" ="Banaskantha",
"Leh(Ladakh)" ="Leh",
"Lahul & Spiti" ="Lahul And Spiti",
"Pashchimi_Singhbhum" ="Pashchimi Singhbhum",
"Purbi_Singhbhum" ="Purbi Singhbhum", 
"Saraikela_Kharsawan" ="Saraikela Kharsawan", 
"Dhar " ="Dhar", 
"North District" ="North  District",
"The_Nilgiris" ="The Nilgiris", 
"Kanshiram_Nagar" ="Kanshiram Nagar",
"Sant_Kabir_Nagar" ="Sant Kabir Nagar",
"Sant_Ravidas_Nagar_(Bhadohi)" ="Sant Ravidas Nagar (Bhadohi)",
"Siddharthnagar" ="Siddharth Nagar",
"Tehri_Garhwal" ="Tehri Garhwal",
"Udham_Singh_Nagar" ="Udham Singh Nagar",
"Dakshin_Dinajpur" ="Dakshin Dinajpur", 
"Koch_Bihar " ="Koch Bihar",
"North_Twenty_Four_Parganas" ="North Twenty Four Parganas",
"Paschim_Medinipur" ="Paschim Medinipur",
"Purba_Medinipur" ="Purba Medinipur", 
"South_Twenty_Four_Parganas" ="South Twenty Four Parganas", 
"Uttar_Dinajpur" ="Uttar Dinajpur"))

#the 41 conflict obs that remain are only metropolitan areas/ towns (with no land use data), other districts in states with no land use/ dde jure land tenure data from census 2011 (e.g.: Arunachal Pradesh) 

#District summary of variables of interest
district_summary = dat.summary%>%group_by(District_Name, State)%>%summarise_at(.vars =vars(Total_Indiv, Total_Common, Total_Forest, Total_Unaccounted, Total_Population, Total_Household, Total_hectares_for_prop, forests_land, non_agricultural_land, barren_uncultivable_land, pasture_grazing_land, misc_tree_crop_land, culturable_wasteland, fallow_other_than_current_fallow_land, current_fallow_land, net_sown_area, total_irrigated_land, total_unirrigated_land), .funs =c(sum = "sum"))

#Combine this table with the mpi table
mpi_district_summary = merge(district_summary, mpi_district[c("District", "State", "MPI_District", "Number_of_MPI_PeopleDistrict", "Living_standards")], by.x = c("State", "District_Name"), by.y= c("State", "District"))

#find the districts with no match between mpi and dat.summary 
conflict_obs_district = district_summary[!district_summary$District_Name %in% mpi_district_summary$District_Name, ]
#no districts missing
#N= 592 districts

```

This data will be stitched with spatial data at the district level.

```{r message=FALSE, warning=FALSE, include=TRUE }

#district_mpi_climate = mpi_district_summary

#convert Living standards to a 0 to 1 scale 
range01 <- function(x){((x-min(x))/(max(x)-min(x)))}

mpi_district_summary$Living_standards_scaled = range01(mpi_district_summary$Living_standards)

##MPI
#rearrange the districts in descending order of MPI
mpi_district_summary  <- mpi_district_summary [order(mpi_district_summary$Living_standards_scaled, decreasing=TRUE),]

#assign each district a rank
mpi_district_summary= mpi_district_summary%>% mutate(MPI_rank = match(Living_standards_scaled, unique(Living_standards_scaled)))

```

Write csv to be used in script 03

```{r}

#write csv 
write.csv(district_mpi_climate, "District-MPI-Combined.csv")

```
