==> 06-District-MPI-Mapped.Rmd <==
---
title: "People-centric restoration: Mapping variables to districts"
author: "Pooja Choksi"
date: '2022-06-22'
output: html_document
---

COMBINING DISTRICT LEVEL LAND USE INFORMATION WITH SPATIAL DATA

```{r setup, include=FALSE}

library(data.table)
library(dplyr)
library(tidyr)
library(sf)
library(rgdal)
library(tools)
library(sp)
library(ggplot2)

```

Read in the spatial data to be combined with the tables from script 02.

```{r message=FALSE, warning=FALSE, include=TRUE}

#India district level spatial data
districts = readOGR("C:\\Users\\pooja\\Documents\\GitHub\\Land-Tenure-Restoration-Potential\\Land-Tenure-Restoration-Potential\\Census2011-District-Shapefiles\\IND_adm2.shp")

districts_as_df = as.data.frame(districts)

#alternate file
district_mpi_summary = read.csv("District-MPI-Combined.csv")

```

CLEAN THE DISTRCIT SPATIAL DATA TO MATCH THE DISTRICT LEVEL MPI AND OTHER SOCIO- ECONOMIC DATA 

```{r message=FALSE, warning=FALSE, include=TRUE}

#first check if all state names match 
state_names_missing =  setdiff(districts$NAME_1, district_mpi_summary$State)   # elements in first NOT in second

state_names_missing

#recode the state names to match the spatial data 
district_mpi_summary = district_mpi_summary%>%dplyr::mutate(State = recode(State, "Jammu And Kashmir" ="Jammu and Kashmir", 
"Andaman And Nicobar Islands" = "Andaman and Nicobar", 
"Dadra And Nagar Haveli" = "Dadra and Nagar Haveli"))

state_names_missing 
#still missing districts are Union territories for which we don't have climate vulnerability data

#Check for district names that don't match
district_names_missing =  setdiff(district_mpi_summary$District_Name, districts$NAME_2)   # elements in first NOT in second
  
#to compare names, create a dataframe of the districts shapefile
districts_df <- as(districts, "data.frame")

#recode the district names to match the spatial data 
district_mpi_climate_summary = district_mpi_climate_summary%>%dplyr::mutate(District_Name = recode(District_Name, 
#"Kanker" = "Uttar Bastar Kanker",
# "Sonepur" ="Subarnapur",
"Baudh" ="Bauda",
"North & Middle Andaman" = "North and Middle Andaman",
"Nicobars" = "Nicobar Islands",
"Dadra & Nagar Havel" = "Dadra and Nagar Haveli",
"East District" = "East Sikkim",
"West District" = "West Sikkim",
"North District" = "North Sikkim",
"South District" = "South Sikkim",
# "Shopian" ="Shupiyan",
# "Kamrup Metro" = "Kamrup Metropolitan",
# "Deogarh" ="Debagarh",
"Janjgir - Champa" = "Janjgir-Champa",
# "North_East" = "North East",
# "North_West" = "North West",
# "South_West" = "South West",
"Panchmahal" = "Panch Mahals",
"Sabarkantha" = "Sabar Kantha",
"Banaskantha" = "Banas Kantha",
"Leh" = "Leh (Ladakh)", 
"Lahul And Spiti" ="Lahul & Spiti",
# "West Singhbhum" = "Pashchimi Singhbhum",
# "West Singhbhum" = "Purbi Singhbhum",   
"Saraikela Kharsawan" = "Saraikela-kharsawan", 
# "Dhar " = "Dhar", 
"Kaimur (Bhabua)" = "Kaimur",
# "North  District"= "North District",
# "The_Nilgiris" = "The Nilgiris", 
"Kanshiram Nagar" = "Kasganj", #new name 
# "Sant Kabeer Nagar" = "Sant Kabir Nagar",
"Sant Ravidas Nagar (Bhadohi)" = "Sant Ravi Das Nagar",
# "Dinajpur Dakshin" = "Dakshin Dinajpur", 
# "Coochbehar" = "Koch Bihar",
"Dohad" = "Dahod",
"North Twenty Four Parganas" = "North 24 Parganas",
"Paschim Medinipur" = "Pashchim Medinipur",
# "Medinipur East" = "Purba Medinipur", 
"South Twenty Four Parganas" = "South 24 Parganas",
"Baramula" = "Baramulla",
"Punch" = "Poonch",
# "Ferozepur" ="Firozpur",
# "Jalandar" = "Jalandhar",
# "Sri Muktsar Sahib" ="Muktsar",
# "S.A.S Nagar" = "Sahibzada Ajit Singh Nagar",
# "Rudra Prayag" = "Rudraprayag",
# "Udam Singh Nagar"= "Udham Singh Nagar",
# "Dinajpur Uttar" = "Uttar Dinajpur", 
"Mahrajganj" = "Maharajganj",
"Virudhunagar" ="Virudunagar",
# "North (Sikkim)" = "North District",
# "South (Sikkim)" = "South District",
# "Saran(chhapra)" = "Saran",
"Chamarajanagar" ="Chamrajnagar",
"Sri Potti Sriramulu Nellore" = "Nellore",
# "Khamman" = "Khammam",
# "Raigad" = "Raigarh",
# "Koderma" = "Kodarma",
# "Gurugram" ="Gurgaon",
# "Purbi Champaran" = "Purba Champaran", 
"Kabirdham" = "Kabeerdham",
"Korea (Koriya)" = "Koriya",
# "Dang" ="The Dangs",
"Bara Banki" = "Barabanki",
"Chikkaballapura" ="Chikballapura",
# "Ballari" = "Bellary",
# "Davangere" ="Davanagere",
# "Kalaburagi" ="Gulbarga",
# "Mysuru" ="Mysore",
# "Shimoga" = "Shivamogga",
# "Tumkur" = "Tumakuru",
# "Beed" ="Bid",
# "Buldhana" = "Buldana",
"Papumpare" = "Papum Pare",
# "Gondia" = "Gondiya",
# "Chittorgarh" = "Chittaurgarh",
# "Jalore" ="Jalor",
# "Jhunjhunu" ="Jhunjhunun",
# "Kanchipuram" ="Kancheepuram",
# "Villupuram" ="Viluppuram",
# "Mahabubnagar" = "Mahbubnagar",
# "Prayagraj" = "Allahabad",
# "Ayodhya" = "Faizabad",
# "Kushi Nagar" = "Kushinagar",
"Kheri" = "Lakhimpur Kheri",
# "Haridwar" = "Hardwar",
# "Uttar Kashi" ="Uttarkashi",
# "Barddhaman " = "Barddhaman",
# "Darjeeling" = "Darjiling",
# "Hooghly" = "Hugli",
# "North & Middle Andaman" = "North And Middle Andaman",
"Khandwa (East Nimar)" = "East Nimar", 
# "Ahmednagar" = "Ahmadnagar", 
# "Bagalkote" ="Bagalkot", 
# "Bandipora" ="Bandipore", 
# "Bengaluru Rural" = "Bangalore Rural",
# "Bengaluru Urban" = "Bangalore",
# "Belagavi" = "Belgaum",
# "Chikkamagaluru" = "Chikmagalur", 
# "South Andaman" = "South Andamans", 
# "Visakhapatanam" = "Visakhapatnam",
#  "Purulia" ="Puruliya",
# "Nuh" ="Mewat",
"Mahamaya Nagar" = "Hathras",
# "Narsinghpur" ="Narsimhapur",
# "Marigaon" = "Morigaon", 
# "Tuticorin" = "Thoothukkudi",
"Shrawasti" = "Shravasti",
# "Siddharthnagar" = "Siddharth Nagar",
# "Budgam" ="Badgam", 
"Khargone (West Nimar)" = "West Nimar", 
# "Nabarangpur" ="Nabarangapur",
# "Sahebganj" ="Sahibganj",
"Rangareddy" = "Ranga Reddy",
# "Howrah" = "Haora",
"Dakshin Bastar Dantewada" = "Dantewada",
"Nagapattinam" = "Nagappattinam",
# "Dholpur" ="Dhaulpur",
# "Tehri_Garhwal" = "Tehri Garhwal",
"Gadchiroli" = "Garhchiroli"))

#Check for district names that don't match
district_names_missing =  setdiff(district_mpi_climate_summary$District_Name,
districts$NAME_2)   # elements in first NOT in second

#9 districts missing becuase they're not in the dataset of land use data

```

COMBINE THE SPATIAL DATA WITH THE DISTRICT-MPI- CLIMATE DATA

```{r message=FALSE, warning=FALSE, include=FALSE}

#Combine the spatial data with district level data
districts_merged <-  sp::merge(districts, district_mpi_climate_summary, by.x = c("NAME_2", "NAME_1"),by.y = c("District_Name", "State"),duplicateGeoms = T)

#only MPI and land use data
writeOGR(obj = districts_merged, layer = "districts", dsn= "District-MPI-Merged-592districts.shp", driver = "ESRI Shapefile")

```
