==> 08-District-Spatial-Analysis-Biophysical-People-Scale.Rmd <==
---
title: "People-centric restoration: Comparing district level biophysical potential with the people-centric restoration potential"
author: "Pooja Choksi"
date: '2022-06-24'
output: html_document
---

```{r setup, include=FALSE}

library(data.table)
library(dplyr)
library(tidyr)
library(sf)
library(rgdal)
library(raster)
library(tools)
library(sp)
library(ggplot2)
```

READ IN ALL THE RASTERS FROM THE BIOPHYSICAL POTENTIAL STUDIES AND DISTRICT SHAPEFILE FROM PREVIOUS CODE

```{r message=TRUE, warning=TRUE, include=TRUE}

#Using a raster of restoration potential from Strassburg et al. 2020, we clipped India's restoration potential in QGIS 

strassburg = raster("C:\\Users\\pooja\\Documents\\GitHub\\Land-Tenure-Restoration-Potential\\Land-Tenure-Restoration-Potential\\Restoration_Potential_Studies\\India_Clipped\\Strassburg_india_Clipped.tif")

#district shapefile with info on the people-centric variables
districts = readOGR("C:\\Users\\pooja\\Documents\\GitHub\\Land-Tenure-Restoration-Potential\\Land-Tenure-Restoration-Potential\\Census2011-District-Shapefiles\\District-MPI-Merged-592districts.shp")
  
```

EXTRACT MEAN BIOPHYSICAL POTENTIAL FOR EACH DISTRICT

```{r message=FALSE, warning=FALSE, include=TRUE}

#extract mean value of restoration potential of raster for each district
strassburg_extracted = raster::extract(strassburg, districts, fun = mean, na.rm = T, small = T, sp=TRUE)

strassburg_extracted_df <- as(strassburg_extracted, "data.frame")

potential_extracted = strassburg_extracted_df

#remove the districts which don't have a value for MPI from the larger districts file and potential extracted
potential_extracted= potential_extracted[!is.na(potential_extracted$Lvng_st),]

potential_extracted= potential_extracted[!is.na(potential_extracted$Strassburg_India_Clipped),]

#3. Strassburg
#scale the strassburg potential % from 0 to 1 
#since the Strassburg paper has a scale of 0-20, we need to first rescale it to a 0 to 100 and then divide to scale it from 0 to 1
library(scales)
potential_extracted$Strassburg_Scaled_100 <-  rescale(potential_extracted$Strassburg_India_Clipped, from = c(0, 20), to = c(0, 100), na,rm=T)

potential_extracted$Strassburg_Scaled_final <- (potential_extracted$Strassburg_Scaled_100/100)

#merge the extracted data with districts shapefile

districts_final = sp::merge(districts, potential_extracted[c("Strassburg_India_Clipped", 
"Strassburg_Scaled_final", "NAME_1", "NAME_2")], by.x = c("NAME_1","NAME_2"), by.y= c("NAME_1", "NAME_2"), duplicateGeoms = T)

#MPI, Strassburg restoration potential and land use file merged for mapping purpose (Figure 1 and 2)
writeOGR(districts_final, layer = "districts_final", dsn= "District-MPI-Merged-Biophys.shp", driver = "ESRI Shapefile")

#also create a dataframe from the shapefile to use for analysis below 
districts_df_final <- as(districts_final, "data.frame")

write.csv(districts_df_final, "District-MPI-Biophys-Combined.csv")

```
