==> 09-Figures-Tables-Maps-in-Manuscript.Rmd <==
---
title: "People-centric restoration: Figures and tables in manuscript"
author: "Pooja Choksi"
date: '2022-06-27'
output: html_document
---

```{r setup, include=FALSE}
library(ggplot2)
library(forcats)
library(dplyr)
library(qcc)
library(gridExtra)
library(viridis)
#get color blind friendly colors 
library(pals)
library(rgdal)
library(ggpubr)
library(ggpmisc)

```


Read in all the data from previous scripts that we need for plots and tables.

```{r message = FALSE, warning=FALSE, include=TRUE}

#For all questions, we need this data
districts_info = read.csv("District-MPI-Biophys-Combined.csv")

#list of aspirational districts from Phase 1
aspirational_districts = read.csv("Aspirational-Districts-Phase1.csv")

#districts shp file with MPI and land use data 
districts = readOGR("C:\\Users\\pooja\\Documents\\GitHub\\Land-Tenure-Restoration-Potential\\Land-Tenure-Restoration-Potential\\Census2011-District-Shapefiles\\District-MPI-Merged-Biophys.shp")
```


#CALCULATE PROPRTION OF LAND TENURE IN EACH DISTRICT
```{r}

#calculate prop of land tenures
districts_info = districts_info%>%mutate(Prop_Indiv = (Ttl_In_/(Ttl_In_+Ttl_Cm_+Ttl_Fr_+Ttl_Un_))*100,
Prop_Common = (Ttl_Cm_/(Ttl_In_+Ttl_Cm_+Ttl_Fr_+Ttl_Un_))*100,
Prop_Unaccounted =(Ttl_Un_/(Ttl_In_+Ttl_Cm_+Ttl_Fr_+Ttl_Un_))*100,
Prop_Forest = (Ttl_Fr_/(Ttl_In_+Ttl_Cm_+Ttl_Fr_+Ttl_Un_))*100)

```

#COMBINE ASPIRATIONAL DISTRICTS TO DISTRICTS_INFO

```{r message=FALSE, warning=FALSE, include=TRUE}

#Make a variable for the presence of aspirational district
aspirational_districts$Aspirational_District = "Aspirational_district"

#recode districts to match with the main dataset 
aspirational_districts = aspirational_districts%>%dplyr::mutate(District.name = recode(District.name, 
"West singhbhum" ="Pashchimi Singhbhum", 
"Chatttarpur" = "Chhatarpur",
"East nimar" = "East Nimar",
"Gadchiroli" = "Garhchiroli",
"Nuapara" = "Nuapada",
"Ferozepur" = "Firozpur",
"Dholpur" = "Dhaulpur",
"Virudhunagar" = "Virudunagar",
"Haridwar" = "Hardwar",
"Udham singh nagar" = "Udham Singh Nagar",
"Dakshin dinajpur" = "Dakshin Dinajpur"))

#combine with districts_info

#Combine aspiration districts with the main dataset
districts_info = merge(districts_info, aspirational_districts[c("District.name", "S.NO.", "Aspirational_District")], by.x = c("NAME_2", "NAME_1"), 
              by.y =c("District.name", "S.NO."), all.x=T)

```

## SYNERGIES AND TRADE-OFFS BETWEEN SEVERAL FACTORS

The chunk of code below creates Figure 2(a) scatterplot where each district's biophysical potential and people-centric restoration opportunity scale is plotted. 

```{r message=FALSE, warning=FALSE, include=TRUE}
#All the land tenures together
#first categorize the three variables

#Remove the duplicates 
districts_info = districts_info [!duplicated(districts_info[c("NAME_1","NAME_2")]),]

districts_info$tenure_max  = apply(districts_info[c("Ttl_Cm_", "Ttl_In_", "Ttl_Fr_")], 1, max)

districts_info = districts_info%>%mutate(tenure_max_cat = 
 case_when(tenure_max == Ttl_In_ ~ "Individual",
           tenure_max == Ttl_Cm_ ~ "Common",
           tenure_max == Ttl_Fr_ ~ "Forest", 
           TRUE ~ "NA"))

#Poverty v/s biophysical potential 

tenure_category = c("#999933","#117733", "#E69F00")

biophys_people_scale = ggplot(districts_info) + geom_jitter(data= districts_info, aes(x = Strassburg_Scaled_final, y= Lvng_s_, color= tenure_max_cat), size=3, alpha = 0.6)+theme_bw()+scale_color_manual(values = tenure_category, labels = c("Non forest commons", "Forest commons", "Private"))+ scale_y_continuous(limits = c(0,1), name = "Poverty (Living standards)")+ scale_x_continuous(limits =c(0,1), name="Biophysical restoration potential")+ geom_vline(xintercept =0.74401004, linetype="dashed") + geom_hline(yintercept = 0.7116883, linetype="dashed")+ labs(color = "Dominant land tenure", shape = "Dominant tenure regime")+ theme(legend.position = "bottom")

biophys_people_scale

ggsave(filename = "Fig-2-Biophys-Pov", device = "jpeg", dpi = 300, width = 1900, height = 1450, units = "px")

```

#AGREEMENT BETWEEN THE DIFFERENT SCALES WITHIN DISTRICTS PERCENTILES 

This chunk of code uses the exact cut offs for the percentiles to see the synergy between districts

```{r message=FALSE, warning=FALSE, include=TRUE}

#Percentiles for Strassburg dataset
#remove any NAs in the data
districts_info= districts_info[!is.na(districts_info$Strassburg_Scaled_final),]

quantile(districts_info$Strassburg_Scaled_final,  probs = seq(0, 1,1/10))

districts_info = districts_info%>%mutate(Strassburg_Percentile_Upper = case_when(Strassburg_Scaled_final<=0.4165800 ~ 10,
Strassburg_Scaled_final<=0.5529359 ~ 20,
Strassburg_Scaled_final<=0.6591955 ~30,
Strassburg_Scaled_final<=0.7021665 ~ 40,
Strassburg_Scaled_final<=0.7420543 ~ 50,
Strassburg_Scaled_final<=0.7735714 ~ 60,
Strassburg_Scaled_final<=0.8034219 ~ 70,
Strassburg_Scaled_final<=0.8319207 ~ 80,
Strassburg_Scaled_final<=0.8688850 ~ 90,
Strassburg_Scaled_final<=0.9565341 ~ 100,
TRUE ~ 0))

#Percentiles for living standards of MPI 
quantile(districts_info$Lvng_s_,  probs = seq(0, 1,1/10))

districts_info = districts_info%>%mutate(Poverty_Percentile_Upper = case_when(
Lvng_s_<=0.47341772 ~ 10,
Lvng_s_<=0.59392405 ~ 20,
Lvng_s_<=0.64810127 ~ 30,
Lvng_s_<=0.67898734 ~ 40,
Lvng_s_<=0.71645570 ~ 50,
Lvng_s_<=0.74177215 ~ 60,
Lvng_s_<=0.77113924 ~ 70,
Lvng_s_<=0.80506329 ~ 80,
Lvng_s_<=0.85620253 ~ 90,
Lvng_s_<=1.00000000 ~ 100,
TRUE ~ 0))

#Agreement between poverty and Strassburg
districts_info = districts_info%>%mutate(Poverty_Strassburg_Agreement = case_when(Poverty_Percentile_Upper >=100 & Strassburg_Percentile_Upper >= 100~ 100,
Poverty_Percentile_Upper >= 90 & Strassburg_Percentile_Upper >=90~ 90,
Poverty_Percentile_Upper >= 80 & Strassburg_Percentile_Upper >=80~ 80,
Poverty_Percentile_Upper >= 70 & Strassburg_Percentile_Upper >=70~ 70,
Poverty_Percentile_Upper >= 60 & Strassburg_Percentile_Upper >=60~ 60,
#Poverty_Percentile_Upper >= 50 & Strassburg_Percentile_Upper >=50~ 50,
Poverty_Percentile_Upper<= 50 & Strassburg_Percentile_Upper<= 50 ~ 1, Poverty_Percentile_Upper<= 50 & Strassburg_Percentile_Upper>= 50 ~ 2, Poverty_Percentile_Upper>= 50 & Strassburg_Percentile_Upper<= 50 ~ 3, TRUE ~ 0))

#Agreement using quartiles (for map)
quantile(districts_info$Lvng_s_,  probs = seq(0, 1,1/4))

quantile(districts_info$Strassburg_Scaled_final,  probs = seq(0, 1,1/4))

districts_info = districts_info%>%mutate(Poverty_Strassburg_Quartile_Upper = case_when(Lvng_s_<=0.62151899 & Strassburg_Scaled_final <=0.6159363~ 25,
Lvng_s_<=0.71645570 & Strassburg_Scaled_final<=0.7420543 ~ 50,
Lvng_s_<=0.79113925 & Strassburg_Scaled_final<=0.8184396~ 75,
Lvng_s_<=1.0000000 & Strassburg_Scaled_final<=0.9565341~ 100,
TRUE~ 0))

```

#SUPP INFO TABLES OF TOP DISTRICTS AS PER EACH SCALE AND AGREEMENT AND DISAGREEMENT IN DISTRICTS

This chunk of code is used to produce tables SI 1,2,3____

```{r message=TRUE, warning=TRUE, include=FALSE}

#S4: Table of aspirational districts in the top people-centric restoration scale 
aspirational_poverty = districts_info%>%filter(Poverty_Percentile_Upper == 100 & Aspirational_District == "Aspirational_district")

aspirational = aspirational_poverty%>%group_by(NAME_1, NAME_2,Lvng_s_)%>%count()

write.csv(aspirational, "TableS4-Aspirational-Districts-90thPer.csv")

#Poverty and Strassburg agreement percentiles
agreement_poverty_strassburg = districts_info%>%group_by(Poverty_Strassburg_Agreement)%>%count()

```

#REPORT THE DISTRICTS IN THE TOP AND BOTTOM PERCENTILES AND QUANTILES (AND LAND TENURES IN THE QUANTILES)

This chunk of code is used to generate the numbers reported in the main text of the paper. Refer to each table created in this chunk of code to compare with the numbers reported in paper

```{r message = FALSE, warning=FALSE, include =TRUE}

#top percentile (higher percentile is higher priority)
percentile_100_people = districts_info%>%filter(People_Percentile_Upper == "100")
percentile_100_states = percentile_100_people%>%group_by(NAME_1)%>%count()
  
#top percentile of poverty 
percentile_100_poverty = districts_info%>%filter(Poverty_Percentile_Upper == "100")

percentile_100_poverty_states = percentile_100_poverty%>%group_by(NAME_1)%>%count()

#report the land tenure in the top 50% of districts which have agreement with strassburg scale

percentile_50_poverty_strassburg = districts_info%>%filter(Poverty_Strassburg_Agreement>=60)

poverty_tenure_max_totals = percentile_50_poverty_strassburg%>%group_by(tenure_max_cat)%>%count()

percentile_50_poverty_strassburg = districts_info%>%filter(Poverty_Strassburg_Agreement>=60)

#agreement between poverty and biophysical potential 
poverty_strass_agreement_100 = districts_info%>%filter(Poverty_Strassburg_Agreement == "100")

poverty_strass_agreement_90 = districts_info%>%filter(Poverty_Strassburg_Agreement == "90")

poverty_strass_agreement_50 = districts_info%>%filter(Poverty_Strassburg_Agreement == "1")

```

#PLOT OF PROPORTION OF LAND TENURES WITHIN THE DISTRICTS ORGANISED BY POVERTY (LVING STANDARDS COMPONENT OF MPI)

This chunk of code was used to create Figure 3. 

```{r message=FALSE, warning=FALSE, include=TRUE}

#Proportion of land available in each land tenure
district_total_plot = districts_info%>%dplyr::group_by(Poverty_Percentile_Upper)%>%
  dplyr::summarise(#total = n(),
                   total_indiv = sum(Ttl_In_), 
                   total_forest = sum(Ttl_Fr_),
                   total_unaccounted = sum(Ttl_Un_),
                   total_common = sum(Ttl_Cm_))

#Bar plot of total hectares per percentile 
district_total_plot_long = reshape2::melt(district_total_plot, id.vars = "Poverty_Percentile_Upper")

#plot
district_stacked_barplot_totals = 
district_total_plot_long%>%
  ggplot(aes(x = factor(Poverty_Percentile_Upper), y = value, fill = variable))+
  geom_col(position = position_stack(reverse = TRUE))+theme_bw()+theme(axis.text.x = element_text(angle = 45, hjust = 1))+
  scale_fill_manual(labels = c("Individual", "Common", "Unaccounted","Forest"),
    values=c("#E69F00", "#999933", "#D55E00","#117733")) + labs(fill = "Land tenure/cover", x = "Percentiles", y = "Proportion of Tenure")+ scale_x_discrete(name = "Percentiles", label = c("<10%", ">10%", ">20%", ">30%", ">40%", ">50%", ">60%", ">70%", ">80%", ">90%"))

#Plot of prop of land tenure
district_prop_plot = district_total_plot%>%mutate(Prop_Indiv = (total_indiv/(total_indiv+total_common+total_forest+total_unaccounted))*100,
                                                 
Prop_Common = (total_common/(total_indiv+total_common+total_forest+total_unaccounted))*100,

Prop_Unaccounted =(total_unaccounted/(total_indiv+total_common+total_forest+total_unaccounted))*100,

Prop_Forest = (total_forest/(total_indiv+total_common+total_forest+total_unaccounted))*100)

#reshape data for plotting
install.packages("reshape2")
library(reshape2)

district_prop_plot_long = district_prop_plot[, -c(2:5)]

district_prop_plot_long = reshape2::melt(district_prop_plot_long, id.vars = "Poverty_Percentile_Upper")

#plot
district_stacked_barplot = 
district_prop_plot_long%>%
  ggplot(aes(x = factor(Poverty_Percentile_Upper), y = value, fill = variable))+
  geom_col(position = position_stack(reverse = TRUE))+theme_bw()+theme(axis.text.x = element_text(angle = 45, hjust = 1), legend.position = "bottom")+
  scale_fill_manual(labels = c("Private", "Nonforest commons", "Unaccounted land","Forest commons"),
    values=c("#E69F00", "#999933", "#D55E00","#117733")) + labs(fill = "Land tenure", x = "Poverty (living standards)", y = "Proportion of Tenure (%)")+ scale_x_discrete(name = "Poverty (living standards)", label = c("<10%", ">10%", ">20%", ">30%", ">40%", ">50%", ">60%", ">70%", ">80%", ">90%")) 

district_stacked_barplot

```

#PLOT OF THE DISTRIBUTION OF EACH LAND TENURE IN THE TOP DISTRICTS AND REPORT % of LAND TENURE IN TOP DISTRICTS

This chunk of code was used to create Fig_ and the values reported in the paper.

```{r message=FALSE, warning=FALSE, include=TRUE}

top_districts = districts_info%>%filter(People_Percentile_Upper ==100)

top_districts = top_districts%>%select("NAME_2", "NAME_1", "Tt_I_",
                                        "Tt_C_", "Tt_F_", "Tt_U_")

#wide to long
top_district_long = reshape(top_districts, direction='long', varying=c('Tt_I_', 'Tt_C_', 'Tt_F_', 'Tt_U_'), timevar= 'Land_Tenure', times=c('Individual', 'Common_NonForest', 'Forest', 'Unaccounted'),
v.names= 'Total_Hectares',
idvar=c('NAME_2', 'NAME_1'))

#plot of the distribution of land tenure in top districts
tenure_topdistricts = ggplot()+geom_histogram(data= top_districts, aes(x=Total_Forest_sum), fill ="#117733", alpha=0.5, binwidth = 300000)+geom_histogram(data = top_districts, aes(x=Total_Common_sum), fill = "#999933", alpha =0.5, binwidth = 300000)+ geom_histogram(data = top_districts, aes(x= Total_Indiv_sum), fill = "#E69F00", alpha= 0.5, binwidth = 300000)+ theme_classic()+ theme(axis.text.x = element_text(angle = 45, hjust = 1))+ labs(x = "Total number of hectares", y = "Freqeuncy")

tenure_topdistricts

#reporting % of land tenure in each district 
tenure_topdistricts_report = top_district_long%>%group_by(Land_Tenure)%>%summarise(hectares_category = sum(Total_Hectares))

tenure_topdistricts_report= tenure_topdistricts_report%>%mutate(tenure_total = colSums(tenure_topdistricts_report[,-1])) 

tenure_topdistricts_report = tenure_topdistricts_report%>%mutate(tenure_prop = (hectares_category/tenure_total)*100)

#how many districts have other tenures as max?
top_districts$tenure_max  = apply(top_districts[c("Tt_C_", "Tt_I_", "Tt_F_")], 1, max)

top_districts_tenure = top_districts%>%mutate(tenure_max_cat = 
 case_when(tenure_max == Tt_I_ ~ "Individual",
           tenure_max == Tt_C_ ~ "Common",
           tenure_max == Tt_F_ ~ "Forest", 
           TRUE ~ "NA"))

#see DF to report the distribution in main manuscript

```

#COMBINE THE DISTRICTS INFO FILE WITH SPATIAL DATA FOR FINAL MAPS

```{r message=FALSE, warning=FALSE,include=TRUE}

#first remove the districts which don't have a value for climate vulnerability from the larger districts file and potential extracted
districts= districts[!is.na(districts$Pp__M),]

districts_final = sp::merge(districts, districts_info[c("Strassburg_India_Clipped", "Strassburg_Scaled_final", "Aspirational_District", "tenure_max", "tenure_max_cat",   "Poverty_Percentile_Upper", "Strassburg_Percentile_Upper", "Prop_Indiv","Prop_Common", "Prop_Unaccounted", "Prop_Forest", "Poverty_Strassburg_Agreement","Poverty_Strassburg_Quartile_Upper", "NAME_1", "NAME_2")], by.x = c("NAME_1","NAME_2"), by.y= c("NAME_1", "NAME_2"), duplicateGeoms = T)

writeOGR(districts_final, layer = "districts_final", dsn= "District-MPI-Climate-Pop-Merged-Biophys-Percentile-Plot.shp", driver = "ESRI Shapefile")

#alternative shapefile 
writeOGR(districts_final, layer = "districts_final", dsn= "District-MPI-Merged-Biophys-Percentile-Plot.shp", driver = "ESRI Shapefile")

#also create a dataframe from the shapefile to use for analysis below
districts_df_final <- as(districts_final, "data.frame")

#save csv
write.csv(districts_df_final, "District-MPI-Biophys-Agreement-Percentiles.csv")

```
